feat(e2e): add integration test for tracing injection
Right now, in order for the test to work, the `axum-logs-endpoint`
example has to be running in the background.
