+++ e2e/Cargo.toml
@@ -9,5 +9,9 @@ lazy_static = "1.4.0"
 portpicker = "0.1.1"
 rand = "0.8.5"
 reqwest = { version = "0.11.10", default-features = false, features = ["blocking", "json", "rustls-tls"] }
+serde = "1.0.180"
+serde_json = "1.0.104"
 which = "4.2.5"
 tempfile = "3.3.0"
+tokio = { version = "1", features = ["full"] }
+shuttle-common-tests = { path = "../common-tests" }
