+++ e2e/tests/integration/main.rs
@@ -2,6 +2,7 @@ pub mod helpers;
 
 pub mod actix_web;
 pub mod axum;
+pub mod custom_tracing_layer;
 pub mod poem;
 pub mod rocket;
 pub mod salvo;
